+++ client/src/types/User.ts
@@ -1,6 +1,7 @@
 import { SocialNetworks } from '@/types/SocialNetworks';
 
 export type User = {
+  id: string;
   name: string;
   email: string;
   profileImage?: string;
